1. Mainly OSX port fixes  -tweak to the structure packing (using pragma pack 1)  -new osx timer routines in new format  -created xcode project for dynamo 2. Redefined LARGE_INTEGER, LONG and DWORD types for all unix 3. Defined generalized types uintXY_t for use by Windows 4. Other code cleanup

+++ src/IOPortTCP.cpp
@@ -920,7 +920,7 @@ DWORD PortTCP::Peek()
 			//otherwise, there are no bytes available, this prevents recv() from blocking
 			bytes_available = 0;
 		}
-		success = ((long)bytes_available != SOCKET_ERROR);
+		success = (bytes_available != SOCKET_ERROR); // was (long)bytes_available ...
 	}
 #if defined(IOMTR_OS_WIN32) || defined(IOMTR_OS_WIN64)
 	else {
